+++ src/AutoDeconstruct.Analysis/PropertySymbolModel.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.CodeAnalysis;
+
+namespace AutoDeconstruct.Analysis;
+
+internal sealed record PropertySymbolModel(
+	string Name,
+	string CamelCaseName,
+	string TypeFullyQualifiedName,
+	Accessibility Accesibility, 
+	Accessibility TypeAccessibility);