+++ src/AutoDeconstruct.Analysis/Shared.cs
@@ -1,0 +1,10 @@
+﻿using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("AutoDeconstruct.Analysis.Tests")]
+
+internal static class Shared
+{
+	internal const string DeconstructName = "Deconstruct";
+	internal const string AutoDeconstructAttributeName = "AutoDeconstruct.AutoDeconstructAttribute";
+	internal const string TargetAutoDeconstructAttributeName = "AutoDeconstruct.TargetAutoDeconstructAttribute";
+}