+++ src/AutoDeconstruct.Analysis/TypeSymbolModelIssue.cs
@@ -1,0 +1,9 @@
+﻿namespace AutoDeconstruct.Analysis;
+
+internal enum TypeSymbolModelIssue
+{
+	None,
+	NoAccessibleProperties,
+	InstanceDeconstructExists,
+	ExtensionsDeconstructExists
+}