+++ src/AutoDeconstruct/HelpUrlBuilder.cs
@@ -1,7 +1,0 @@
-﻿namespace AutoDeconstruct;
-
-internal static class HelpUrlBuilder
-{
-	internal static string Build(string identifier) =>
-		$"https://github.com/JasonBock/AutoDeconstruct/tree/main/docs/{identifier}.md";
-}