+++ src/AutoDeconstruct/Shared.cs
@@ -1,8 +1,3 @@
 ﻿using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("AutoDeconstruct.Tests")]
-
-internal static class Shared
-{
-	internal const string DeconstructName = "Deconstruct";
-}