+++ src/AutoDeconstruct/TypeSymbolModelIssue.cs
@@ -1,9 +1,0 @@
-﻿namespace AutoDeconstruct;
-
-internal enum TypeSymbolModelIssue
-{
-	None,
-	NoAccessibleProperties,
-	InstanceDeconstructExists,
-	ExtensionsDeconstructExists
-}